fixed wrong seed type

+++ src/seeds.c
@@ -1,6 +1,6 @@
 #include "seeds.h"
 
-int seed;
+uint64_t seed;
 
 void
 setSeed(int64_t s)
@@ -12,3 +12,4 @@ nextChar(void){
 seed = (seed * 0x5DEECE66DL + 0xBL) & ((1LL << 48) - 1);
 return (uint8_t)(seed>>40);
 }
+
